+++ examples/demo-rollup/provers/sp1/src/lib.rs
@@ -1,4 +1,4 @@
-use lazy_static::lazy_static;
+use std::sync::LazyLock;
 
 /// Attempt to load the ELF file at the given path. If the file is empty or cannot be read,
 /// a warning will be printed and the default empty vector will be returned.
@@ -13,21 +13,27 @@ fn load_elf(path: &str) -> &'static [u8] {
 
 // Initialize the SP1 guest ELFs. Note: Normally this is done with include_bytes!(PATH_TO_FILE),
 // but because we don't include the guest ELFs in the GitHub build, they may potentially not exist.
-lazy_static! {
-    pub static ref SP1_GUEST_MOCK_ELF: &'static [u8] = load_elf(&format!(
+pub static SP1_GUEST_MOCK_ELF: LazyLock<&'static [u8]> = LazyLock::new(|| {
+    load_elf(&format!(
         "{}/guest-mock/target/elf-compilation/riscv64im-succinct-zkvm-elf/release/sov-demo-prover-guest-mock-sp1",
         env!("CARGO_MANIFEST_DIR")
-    ));
-    pub static ref SP1_GUEST_CELESTIA_ELF: &'static [u8] = load_elf(&format!(
+    ))
+});
+pub static SP1_GUEST_CELESTIA_ELF: LazyLock<&'static [u8]> = LazyLock::new(|| {
+    load_elf(&format!(
         "{}/guest-celestia/target/elf-compilation/riscv64im-succinct-zkvm-elf/release/sov-demo-prover-guest-celestia-sp1",
         env!("CARGO_MANIFEST_DIR")
-    ));
-    pub static ref SP1_GUEST_MOCK_NOMT_ELF: &'static [u8] = load_elf(&format!(
+    ))
+});
+pub static SP1_GUEST_MOCK_NOMT_ELF: LazyLock<&'static [u8]> = LazyLock::new(|| {
+    load_elf(&format!(
         "{}/guest-mock-nomt/target/elf-compilation/riscv64im-succinct-zkvm-elf/release/sov-demo-prover-guest-mock-nomt-sp1",
         env!("CARGO_MANIFEST_DIR")
-    ));
-    pub static ref SP1_GUEST_CELESTIA_NOMT_ELF: &'static [u8] = load_elf(&format!(
+    ))
+});
+pub static SP1_GUEST_CELESTIA_NOMT_ELF: LazyLock<&'static [u8]> = LazyLock::new(|| {
+    load_elf(&format!(
         "{}/guest-celestia-nomt/target/elf-compilation/riscv64im-succinct-zkvm-elf/release/sov-demo-prover-guest-celestia-nomt-sp1",
         env!("CARGO_MANIFEST_DIR")
-    ));
-}
+    ))
+});