+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2026-03-30
+- #2654 Replace `lazy_static` crate with `std::sync::LazyLock`
+
 # 2026-03-24
 - #2626 EVM: Fixes estimateGas value to match what will end up in the receipt of actually executed transaction
 - #2634 Test only changes